requirements-dev.txt
# Development & test dependencies for running CI/tests locally
pytest>=8.3.0,<9.0.0
rdflib>=7.0.0,<8.0.0
pyshacl>=0.26.0,<0.27.0
owlready2>=0.46,<0.47
httpx>=0.27.0,<0.28.0
# UI and agent framework for the interactive agent
gradio>=4.0.0
openai>=1.30.0,<2.0.0
requests>=2.28.0
PyYAML>=6.0
# Optional tooling
ruff>=0.5.0,<0.6.0
